requirements.txt
annotated-types==0.7.0
appdirs==1.4.4
asgiref==3.8.1
asteval==1.0.6
astunparse==1.6.3
attrs==25.1.0
blinker==1.9.0
boto3==1.36.16
botocore==1.36.16
bw-migrations==0.2
bw2calc==2.0.1
bw2data==4.4.2
bw2io==0.9.6
bw2parameters==1.1.0
bw_processing==1.0
cached-property==2.0.1
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
constructive_geometries==1.0
country_converter==1.3
dataflows-tabulator==1.54.3
datapackage==1.15.4
deepdiff==7.0.1
Deprecated==1.2.18
diff-match-patch==20241021
Django==5.1.5
django-import-export==4.3.5
django-unfold==0.47.0
docopt==0.6.2
et_xmlfile==2.0.0
flexcache==0.3
flexparser==0.4
fsspec==2025.2.0
greenlet==3.1.1
idna==3.10
ijson==3.3.0
isodate==0.7.2
jmespath==1.0.1
jsonlines==4.0.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
linear-tsv==1.1.0
llvmlite==0.44.0
loguru==0.7.3
lxml==5.3.1
matrix_utils==0.6
morefs==0.2.2
mrio-common-metadata==0.2.1
mysql-connector-python==9.2.0
numba==0.61.0
numpy==1.26.4
openpyxl==3.1.5
ordered-set==4.1.0
pandas==2.2.3
peewee==3.17.9
pillow==11.1.0
Pint==0.24.4
platformdirs==4.3.6
prettytable==3.14.0
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
pyecospold==4.0.0
pyparsing==3.2.1
PyPrind==2.11.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.2.1
pytz==2025.1
pyxlsb==1.0.10
PyYAML==6.0.2
randonneur==0.6.1
randonneur_data==0.5.4
RapidFuzz==3.12.1
rdflib==7.1.3
referencing==0.36.2
requests==2.32.3
rfc3986==2.0.0
rpds-py==0.22.3
s3transfer==0.11.2
scipy==1.15.1
six==1.17.0
snowflake-id==1.0.2
SPARQLWrapper==2.0.0
sparse==0.15.5
SQLAlchemy==2.0.38
sqlparse==0.5.3
stats_arrays==0.7
structlog==25.1.0
tableschema==1.21.0
tablib==3.8.0
toolz==1.0.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
unfold==1.2.3
unicodecsv==0.14.1
urllib3==2.3.0
voluptuous==0.15.2
wcwidth==0.2.13
wheel==0.45.1
wrapt==1.17.2
wurst==0.4
xlrd==2.0.1
XlsxWriter==3.2.2
